1_Purpose_Built_AI_Services_Demo_Apps/Insurance_Claim_Submission/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
boto3==1.29.7
python-dotenv==1.0.0
pydantic>=2.10.0

